+++ cms/src/features/posts/Posts.tsx
@@ -1,12 +1,18 @@
+import { useState } from "react";
 import PostForm from "./components/PostForm";
+import Post from "./components/Post";
 
 const Posts = () => {
-  // Get posts from database
+  // Get user's posts from database?
+
+  const [posts, setPosts] = useState([]);
+  const postList = posts.length > 0 && posts.map((post) => <Post />);
 
   return (
     <>
       <h2>Posts</h2>
       <PostForm />
+      {postList}
     </>
   );
 };