+++ cms/src/layouts/ProtectedLayout.tsx
@@ -1,9 +1,8 @@
 import React, { useEffect } from "react";
 import { Outlet, useLocation, useNavigate } from "react-router";
 import { useAuth } from "../hooks/useAuth";
-// import axiosDefault from "../config/axios.config";
 // import { usePrevLocation } from "../hooks/usePrevLocation";
-// import useRefreshToken from "../hooks/useRefreshToken";
+
 import useAxiosPrivate from "../hooks/useAxiosPrivate";
 
 const ProtectedLayout: React.FC<{ children?: React.ReactNode }> = ({
@@ -13,7 +12,6 @@ const ProtectedLayout: React.FC<{ children?: React.ReactNode }> = ({
   const { accessToken, setAccessToken } = useAuth();
   const location = useLocation();
   const axiosPrivate = useAxiosPrivate();
-  // const refresh = useRefreshToken();
   // console.log(prevLocation);
   // if (!accessToken)
   //   return <Navigate to="/login" state={{ from: location }} replace />;
@@ -26,12 +24,7 @@ const ProtectedLayout: React.FC<{ children?: React.ReactNode }> = ({
 
     const authorize = async () => {
       try {
-        /* const response = await axiosDefault.get("/auth", {
-          withCredentials: true,
-        }); */
         await axiosPrivate.get("/auth");
-
-        // setAccessToken(response.data.data.accessToken);
       } catch (err) {
         console.log("authorize err caught");
         console.log(location);